Complete Phase 4: the face pipeline
731 tests passing (611 unit + 120 parity), 2 skipped, build clean.

Ports face_detector (4 families), face_landmarker, face_recognizer,
face_classifier, face_masker, face_selector, face_tracker, face_store,
face_creator and content_analyser.

Verified against real ONNX inference rather than reimplementations:
models download via pre_check(), so the C# pipeline is compared against
the actual Python pipeline running the same weights on the same image.
Every model-input tensor matched Python exactly - the check that
matters, since matching inputs mean any output difference is ORT's own
while differing inputs mean a preprocessing bug.

One measured divergence is not a port defect: OpenCvSharp's OpenCV
build and opencv-python resolve bilinear interpolation differently, so
WarpAffine differs by up to 2/255 on ~9% of pixels given identical
matrices. Isolated by dedicated tests; only the two assertions it
cascades into are loosened, and every discrete output stays tight.

The content analyser NSFW gate is ported complete with no bypass, plus
an integrity check whose limits are documented rather than overstated.

+++ tests/FaceFusion.UnitTests/ModelHelperTests.cs
@@ -90,7 +90,7 @@ public sealed class ModelHelperTests
             // float16 has ~3 decimal digits of precision; float32/int/double fixtures
             // round-trip exactly, so a tight-but-nonzero tolerance covers every case
             // uniformly without special-casing dtypes here.
-            Assert.Equal(expectedValues[i], actualValues[i], precision: 3);
+            Assert.Equal((double)expectedValues[i], (double)actualValues[i], precision: 3);
         }
     }
